feat: improved asset pipeline (#190)
* Removing previous project files
Adding renderer project

* changes to project settings

* big overhaul part 1

* fixed issue with static meshes using wrong shader
Added functions for animations
set up animation example using entity system

* removed assets folder

* initializing value to 0

* small changes and new assets

* made animator struct to handle animations for characters
set up an example of animations being played and swapped with the animator
added a default character mesh, animations and texture

* removed old renderer functions that are no longer of use
this included some hard coded things that were only for show

* renamed some asset files
added scaling function for matrix
added a different model being animated in the scene for showcase

* small value changes

* merge changes from main

* finished integrating renderer changes and cleaned up some of the renderer code

* adding some assets into the world

* preparing to load textures from list

* fixed strings used for texture loading

* changed the way textures, meshes and animations are loaded and used in game

* removed unused assets from repo

+++ src/cpp/tile.cpp
@@ -18,8 +18,8 @@ bool tile::line_of_sight(tile* const to,
   int y0 = this->row;
   int y1 = to->row;
 
-  int deltax = std::fabs(x1 - x0);
-  int deltay = std::fabs(y1 - y0);
+  int deltax = static_cast<int>(std::fabs(x1 - x0));
+  int deltay = static_cast<int>(std::fabs(y1 - y0));
 
   int total_tiles = (deltax > deltay) ? deltax : deltay;
   for (size_t i = 0; i < total_tiles; ++i) {
